Write ENVI class lookup so classification files open in ENVI
ENVI requires classes, class names, and RGB class lookup to treat
a file as a classification image. The previous header omitted the
lookup table (and used 16-bit data). Write unsigned 8-bit maps with
a class lookup matching the overlay palette.

+++ tests/test_identification_hbm_adapt.py
@@ -105,7 +105,7 @@ class DisplayFallbackTests(unittest.TestCase):
             )
             header = path.with_suffix(".hdr").read_text(encoding="ascii")
             self.assertIn("classes = 3", header)
-            data = np.fromfile(path, dtype="<i2").reshape(2, 2)
+            data = np.fromfile(path, dtype=np.uint8).reshape(2, 2)
             self.assertEqual(int(data[0, 0]), 1)
             self.assertEqual(int(data[0, 1]), 2)
             self.assertGreater(int(np.sum(data == 1)), 0)
